[C API] Support for executing models (#8034)
Add initial support for `execute` on an async model in the C API.
Internally, `execute` allocates the output tensors that the user is
responsible for releasing the memory for via `M_freeAsyncTensorArray`.

This involves creating a few new DriverAPI functions for converting to
framework specific tensors from `GenericML::Tensor`, releasing the
memory for the framework specific tensor types, and running a model with
a set of inputs. These are all implemented by both TF and tfl.

Note:
- In implementing `execute`, we somehow broke support for gracefully
handling invalid model paths. We'll fix that in a follow-up since we're
more interested in full execute support in a happy path case for now.
For now, remove that part of the `api-test.mlir` for TFL e2e test.
- It would be nice to extend the e2e test to show the C API and `mt`
both produce the same results given the same inputs, but we're not quite
there yet. Right now, it's a manual spot check that holds true for the
MNIST model in the TFL e2e case.
- This functionality will get leveraged soon e2e in the Python APIs
MAX_INTEGRATION_TESTS_REV_ID: ba58f985fe3914d46f307efa296ff3dff83d28fd