[CI] Skip downloading Python if in Dockerfile (#15162)
Now that Python 3.8 is in the Dockerfile as of

[Internal link]
which is used in CI, we don't need to download it in our `setup-python`
action. Skip the download if the `version` argument specified to this
action is `3.8`.

Additionally, stop specifying `--root-user-action=ignore` since `pip`
from the Dockerfile will have the right permissions and not be `root`.

Note:
- It would be nice to refactor this to be more generic/skip the Python
download if the version is specified and not the default, but that's not
required right this second.
MAX_INTEGRATION_TESTS_REV_ID: 8b4837c30133f83df106263a6cac99f17ce8d5f8