[mojo] Restructure linking packages at the pre-elaboration phase (#29758)
Some users want to link or use pre-elaborated IR in a kind of LTO
phase, where everything is compiled together. The current way of
just grabbing the pre-elaboration module isn't actually desirable,
because it still thinks all of the package APIs want to be exported
(which results in a lot of compilation/weirdness). This PR adds
some utilities to allow pulling in the pre-elaboration IR and dropping
the implicit exports. This allows for removing the last use of
--parsing-stdlib from the custom ops tests.

See #29010

MAX_INTEGRATION_TESTS_REV_ID: 16fc25bb9ba9a10d7072da5c3aae6a857809ca86