[Serving] [Telemetry] Add missing check for request package (#41340)
Custom metrics tests requires the `request` python package to query the
end-point and validate that the metrics reported correspond to the
operations performed during the test. The earlier version of the PR
missed the lit-cfg update to check if the `request` module is available
or not.

Signed-off-by: Rashid Kaleem <rashid@modular.com>
MAX_INTEGRATION_TESTS_REV_ID: d64b7df671d3b225b8621737ae233e9c17056fe4