[MOGG] Option to build kernels with debuginfo at compiletime (#44803)
Adds a flag `--with-debuginfo` to build kernels with debuginfo. Under
the hood, this flag loads a .mlirbc file of the kernels that has been
compiled with debuginfo. This mlir bytecode is not embedded because it
is too large for the compiler to handle.

Fixes GRA-684
Fixes GRA-681

MAX_INTEGRATION_TESTS_REV_ID: 3d361f75e3de04304295514a7bcb526609129425