[lit] Delete has_gpu (#46258)
With bazel we have to annotate targets has whether or not they require a
a GPU, so that bazel knows where to execute them in the remote execution
case. Because of this for lit we always add the gpu related features in
bazel to make sure we don't accidentally run gpu tests on a CPU machine
(which would imply we had a configuration error).

The has_gpu function that we were using for this conflicts with that
setup, and generally adds unnecessary work. This became a real problem
in python tests where we didn't have this same safety net and running
this function could fail silently, leading us to not run GPU tests ever.

We now no longer check this in tests, and rely on bazel to execute these
(or skip them) based on the target hardware.

To avoid silently breaking cmake we now always add these features in
cmake too, which could mean we run tests that shouldn't be run on some
machines, but we'd prefer they fail loudly than not run for folks are
still trying to test GPU things with cmake. If needed we can gate this
behind an env var or something, but since cmake's days are numbered I
didn't do this up front.

MAX_INTEGRATION_TESTS_REV_ID: 94290df159157172666686718f7a60903e61e525