[EngineAPI] Remove drop_counter custom op (#47168)
This was intended to deallocate, back when I was originally told that a
`mo.opaque` was just a `void*` and needed to be deallocated manually
with a user op. Now we know that they do carry a destructor, so no
explicit drop operation is needed. We may still want to have a way for
opaques to be consumed via a graph, but that wasn't what I was trying to
test here -- I just wanted to avoid a memory leak.

MAX_INTEGRATION_TESTS_REV_ID: 1066f7f0792d6da032ff788096b82926e4f93b77