[Pipelines] use builtin layer norm kernel for `torch_layer_norm` (#50313)
Fix the implementation of `torch_layer_norm` to use the builtin kernel.
This fixes the problem that the standard deviation needs to be computed
with `correction=0` and hopefully avoids known problems with torch.std
not returning consistent results depending on input shape.

The test can't be turned on yet because the CPU kernel needs to be
updated to avoid catastrophic cancellation caused by the naive variance
algorithm. I thought setting `max_magnitude` in the test might help but
the problem still occurs but with a smaller magnitude.

Related to KERN-1145.

MAX_INTEGRATION_TESTS_REV_ID: 7969ccafee16324d9852d19613915ad5cba74ca3