[API][GenericML] Plumb inputDevices and outputDevices (#50850)
This PR is the vertical plumbing that connects the `input_devices` and
`output_devices` attributes populated in MEF with the users in the
Python API. The different model framework executors are not supported,
and are un-implemented, while MGPModel plumbs the input and output
devices
through the engine so users can refer to LoadedModels for the expected
deviceRef for each input and output.

This also adds a new `M_get*DeviceRefs` C API for the following
scenarios:
* all devices used by the model
* list of devices corresponding to inputs
* list of devices corresponding to outputs.

Due to this change, update most `M_*execute*` API calls to respect the
input and output devices from the model as the SSOT, while using the
devices from the Engine::Config as a way to error check the devices are
setup properly.

Also adds a python property so the input/output devices can be accessed
from the
compiled Model like input/output metadata.

The devices property is kept the same as it is for now.

Towards AIPIPE-115

MAX_INTEGRATION_TESTS_REV_ID: 7d57d1dee8536a88c44d66d6ae6b273e85eeb91e