[Pipelines] Add pipelines-lm-eval tool (#51321)
This tool runs pipelines in the background, waiting for it to become
available, and then runs lm-eval, gracefully stopping the pipelines
server when it finishes (successfully or not).

We also add an lm-eval wrapper so that we can run lm-eval's normal CLI
from Bazel. In doing so, we also add extra lm-eval installation options
to allow running IFEval over local OpenAI API, which is all we've
successfully tested so far for running lm-eval over pipelines serving.

The argument passing is a little verbose but it is unambiguous and I
expect this will mostly be invoked from CI where the verbosity is no big
deal.

Example usage:

./bazelw run SDK/integration-test/pipelines/python/pipelines-lm-eval --
\
        --pipelines-probe-port 8000 \
        --pipelines-probe-timeout 600 \
        --pipelines-arg=llama3 \
        --pipelines-arg=--quantization-encoding=q4_k \
        --pipelines-arg=--serve \
        --lm-eval-arg=--model=local-chat-completions \
        --lm-eval-arg=--tasks=ifeval \

--lm-eval-arg=--model_args=model=meta-llama/Meta-Llama-3.1-8B-Instruct,base_url=http://localhost:8000/v1/chat/completions,tokenized_requests=False
\
        --lm-eval-arg=--apply_chat_template
MAX_INTEGRATION_TESTS_REV_ID: ad0d7f43e628bfcc4dd4d4180450fdc580167b79