[SDK] Remove test_vision_encoder.py (#53345)
Remove test_vision_encoder.py since the granularity of this test is too
coarse: we should prefer pipeline verification at this level.
To that point, this end-to-end test here wasn't correct (rtol = 20), and
the layer test was correct but the pipelines verification was incorrec
(so something was wrong but this test didn't catch it).
I think this was due to re-implementing MllamaVisionEncoderLayer instead
of importing from HF.
So just remove this test in favour of pipelines verification.
MAX_INTEGRATION_TESTS_REV_ID: 6aa47d45066eef39dde469202fa51b83c75fd906