[Pipelines] Add alternate encodings when loading model (#53410)
With the checkpoint encoding validation updates, it was no longer
possible to run `SmolLM` on CPU with float32 since the repo only
contains bfloat16 weights. We'd still like to run float32 when running
tests on CPU, so this PR allows us to do that.

For Safetensor checkpoints, [the weights are cast to the expected
[Internal link]
For GGUF checkpoints, [an error is raised
[Internal link]

Casting bfloat16 to float32 results in about the same level of text
generation.

Command:
`./bazelw run SDK/public/max-repo/pipelines/python/pipelines -- generate
--max-length 100 --max-cache-batch-size 8
--huggingface-repo-id=HuggingFaceTB/SmolLM2-135M --quantization-encoding
float32`

generated from `--quantization-encoding bfloat16 --use-gpu`:
`I believe the meaning of life is to be found in the fact that we are
all connected to each other. We are all connected to the universe. We
are all connected to each other. We are all connected to each other. We
are all connected to each other. We are all connected to each other. We
are all connected to each other. We are all connected to each other. We
are all connected to each other. We are all connected to each other. We
are all connected to each`

generated from `--quantization-encoding float32`:
`I believe the meaning of life is to be found in the fact that we are
all part of the same family. We are all part of the same family. We are
all part of the same family. We are all part of the same family. We are
all part of the same family. We are all part of the same family. We are
all part of the same family. We are all part of the same family. We are
all part of the same family. We are all part of`

MAX_INTEGRATION_TESTS_REV_ID: 40d2b1dad6dcc53e3c64696c960c4a57803fa610