[Driver] Add an "api" property the Device (#53632)
This "api" property will allow the user to determine if a given device
is programmed with a API. Current possible values are:

- "cpu" for host devices
- "cuda" for NVIDIA GPUs
- "hip" for AMD GPUs

Note that this change does not test this API yet on GPUs, since we don't
have a way for the build/CI system to determine which GPU a test is
running on yet. Once that does exist, we should add GPU tests for this
API as well.

MAX_TESTS_TESTS_ORIG_REV_ID: bd6b2c873ddebacdc5fb1451a5d8c11b4ccb5d40