[bazel] Set env variables in lit tests for all potential tools (#56241)
Previously we only set the environment variables for the tools in the
direct data / tools field. Now we set the known paths for every
potential tool. It's still up to the test to include these in their deps
if they are not otherwise included, but now if you include mojo, you
don't need to manually include CompilerRT / orcrt. This is a bit coarse
in that if you depend on something directly, but get it through a
transitive dep, if that dep ever changes you will have to make the new
connection in the test manually, but that seems worth not having to
include these in the common case.

MAX_INTEGRATION_TESTS_REV_ID: ad8bc0f162307a29295fd630d290f4c7ba7b8b32