[Pipelines] Correct Rope Scaling in Rotary Embedding
Our rope scaling tensor was not actually a scaling factor to multiply
in.
It is instead the inv_freq as defined here:
https://github.com/vllm-project/vllm/blob/3e1d223626499a75edfd3f2ed463197e2c7b387b/vllm/model_executor/layers/rotary_embedding.py#L825
If it is defined, it replaces an entire chunk of the frequency
computation.

This PR completely removes rope scaling related calculations from the
weight adapters.
Instead, it puts the calculations directly in RotaryEmbedding.
Specifically, it adds a new Llama3RotaryEmbedding that applies the
scaling in the graph.

In a follow up, we also should go through and attempt to restrict CI
verification atol and rtol now that this is fixed.

MAX_INTEGRATION_TESTS_REV_ID: 0ea82dd8efbb999d5f8893384f99191a6ccfbc07