[Driver] Add repro test with copying unaligned tensor (bug) (#57605)
This PR adds a test for copying unaligned tensors. Currently, there is a
bug so the results of the test are incorrect (the test is written to
pass in CI - once the bug is fixed the test should be corrected as
well).

Linear ticket :
[Internal link]

MAX_TESTS_TESTS_ORIG_REV_ID: a68ab756eb8805f10ad9af65e3776f14d7417a58