[Pipelines] Add option in generate_llm_logits to set batch size > 1 (#59544)
Adds a `--max-batch-size` flag to `generate_llm_logits` to execute a
model with batch size > 1.

We ran into a bug with Llama4 when trying to run benchmarks with batch
size > 1. This has since been fixed, but it would be nice to be able to
easily run a model with larger batch sizes without having to go through
`serve`.

When testing this with `SmolLM`, I discovered minor logit differences.
The output generated text match, but the fact that the results aren't
numerically equivalent is a bit concerning. Ticket:
[Internal link]

Fixes MODELS-483

MAX_INTEGRATION_TESTS_REV_ID: 51163f489e814c1c6788c832052685ed38f1892e