[KVCache] Slight simplification of prefetch API
Instead of allocating just enough blocks to run CE on a chunked prefix
of the prompt,
we allocate blocks for the entire prompt. We will likely need to
allocate blocks for
the rest of the prompt in the following iteration anyways based on our
chunked prefill
implementation.

This slight change in functionality (that is not even necessary worse),
allows
us to get rid of the `reuse_blocks_from_prefix_cache` method and the
`allocate_new_blocks`
calls in the scheduler and replace it with a single `prefetch` call.

Future work involves prefixing all private methods in `block_manager.py`
with
an underscore.

MAX_INTEGRATION_TESTS_REV_ID: 9cf928a6e93760509212e06e9a6e968081caa241