[SDK] Allow ops to be used directly within torch.compile functions (#61947)
The lazy instantiation of `CustomOpDef`s inside of `CustomOpLibrary` is
not legal within torch.compile. This is due to the process bottoming out
to nanobind functions that torch.compile does not know how to handle.

This change guards that logic behind a `torch.compiler.disable` call.
This implementation has tradeoffs, as it will introduce a graph break
when encountering a use of `op_library.opname` within a torch.compile
function if that operation has not been instantiated already.

The idea solution would be to instantiate all the kernel definitions
ahead of time, but the MojoLibraryAnalysis does not give us the ability
to distinguish user kernels from the other kernels that are incidentally
loaded for all the MO operations.

This is hopefully a short-term workaround.

Closes: GEX-2255
MAX_TESTS_TESTS_ORIG_REV_ID: 4c51215f4fb2ea6d9d39310e76c4ff4ff65ed27f