[SDK] Add TextGenerationRequest for pipeline verify (#62915)
Add a `TextGenerationRequest` for pipeline verification and use it in
`run_text_generation` functions.
A follow up will use this to support multiple different types of text
generation requests in a single "text generation run" for a model:

- Text-only.
- Single-image + text.
- Multi-image + text.

This is desirable because image-text to text models such as InternVL are
expected to support all 3 types of inputs.
And so we should have test coverage of all 3.
In the future we can support video input as well.

MAX_INTEGRATION_TESTS_REV_ID: 15a2f96729e4ae1451dfecb398f589aa8646ee28