[SDK] Error when operation is not found in custom library (#63009)
The current implementation errors in an obscure and difficult to
diagnose way when a custom op library does not contain the requested
operation. This checks for the presence of the operation and raises and
exception if the requested operation cannot be found.

Closes: GEX-2320
MAX_TESTS_TESTS_ORIG_REV_ID: f9867b8498a95cabff1e863c9832f18f5a37c9b3