[Pipelines] Correct Type Hints for Llama Vision
The type hints used in Llama Vision are not consistent with their
construction or use in the model. When moving to Msgpack Serialization,
we recreate the class with the type hints as provided and run into a
series of bugs as their use is no longer compatible with the actual
intended types provided. This steps around this by ensuring the types
are used as intended in the model.

MAX_INTEGRATION_TESTS_REV_ID: 21eaf79a1015b822b134352c3566e9e02372ab84