[SDK] Reland "Generate python bindings for passes" (#63851)
- Defines a `Pass` type for `mlir::Pass`
  - `mlir::Pass` doesn't have typeinfo, so we can't bind it directly
- Supports generating `Pass` binding definitions via tblgen
- Generates passes for `builtin`, `mo`, and `rmo` dialects
  - Supports constructing passes with configured options
- Adds a `lower` function which creates and executes a pass manager on
an input `ModuleOp` for a list of passes
  - Correctly captures and raises exceptions on diagnostics
- Adds tests for
  - Successfully executing a pass
  - Unsuccessfully executing a pass and raising with a diagnostic
  - Constructing passes with and without default options

This was reverted due to blowing up the Python extension module size
from ~18MB to ~150 MB. This gets built 5 times, so increases like this
are easy to accidentally slip in and can be devastating.

This adds some fixes to link in the required dependencies into `libmax`
instead - it's a little hacky and should definitely be cleaned up in the
future.

The Python modules still grew slightly to ~21 MB, but I think this is
reasonable given the additions.

Just to be sure:
[Internal link]

---------

Co-authored-by: Stef Lindall <stef@modular.com>
MAX_TESTS_TESTS_ORIG_REV_ID: 3512d92ca94b49d2b8b3fc7e3f7ce72cd2e1889e