[Pipelines] Add new log probabilities implementation
This new implementation runs as a custom kernel invoked as a custom
operation from within a MAX Graph. The new implementation is not itself
any faster than the current implementation, but it can run on the GPU.
The new implementation currently intentionally matches the existing
implementation in signature where possible, but a future PR will change
the tensor input parameters from `np.ndarray` to `Tensor`, allowing us
to remove a transfer operation that currently takes a lot of time.

Tests are hooked up, but models are not modified to use the new
implementation yet. That will follow.

The implementation approach splits some of the work between Python code
and the kernel. The kernel is responsible for picking out the tokens
that need to be picked out, finding the top-k items as well as the
selected token's input logit (though not sorting those), and computing
logprobs. The Python code is responsible for narrowing down the top-k to
the k actually in the request, doing a final sort, as well as preparing
output indices for use by the kernel.

Performance has not been tuned. The thought is that the problematic
transfer is currently the bottleneck. After that is removed, we should
still see a speed-up even if the kernel itself is not blazing fast.

MAX_INTEGRATION_TESTS_REV_ID: 45ce6ccbe9e9c9f78ea7894ea3f89f0a789f46e8