[Models] Remove manual concatenation in the DeepSeek weight adapter
Previously, the graph compiler had trouble dealing with graphs that
involve a significant number of operations. To work around this
limitation, we concatenated all expert weights in the weight adapter,
eliminating the need to concatenate them within the graph. This PR
removes this workaround and passes the expert weights to the graph
compiler as-is.

MAX_INTEGRATION_TESTS_REV_ID: 36bb8ff203e89e24a8aa075a469151e475dad71c